+++ classes/output/core_renderer_toolbox.php
@@ -36,6 +36,7 @@ use core\output\html_writer;
 use core\output\pix_icon;
 use core\url;
 use core_block\output\block_contents;
+use custom_menu as core_custom_menu;
 use navigation_node;
 use stdClass;
 
@@ -1586,37 +1587,19 @@ trait core_renderer_toolbox {
     }
 
     /**
-     * Compares two course entries against their access time for a user to see which is first.
+     * Returns the page heading menu.
      *
-     * @param stdClass $a A course.
-     * @param stdClass $b A course.
-     *
-     * @return int -1 'a' is first, 1 'b' is first or 0 they are equal.
+     * @return string Markup.
      */
-    protected static function timeaccesscompare($a, $b) {
-        // The timeaccess is lastaccess entry and timestart an enrol entry.
-        if ((!empty($a->timeaccess)) && (!empty($b->timeaccess))) {
-            // Both last access.
-            if ($a->timeaccess == $b->timeaccess) {
-                return 0;
-            }
-            return ($a->timeaccess > $b->timeaccess) ? -1 : 1;
-        } else if ((!empty($a->timestart)) && (!empty($b->timestart))) {
-            // Both enrol.
-            if ($a->timestart == $b->timestart) {
-                return 0;
-            }
-            return ($a->timestart > $b->timestart) ? -1 : 1;
-        }
+    public function page_heading_menu() {
+        $retr = parent::page_heading_menu();
 
-        /* Must be comparing an enrol with a last access.
-           -1 is to say that 'a' comes before 'b'. */
-        if (!empty($a->timestart)) {
-            // If 'a' is the enrol entry.
-            return -1;
+        $localtoolbox = \theme_adaptable\toolbox::get_local_toolbox();
+        if (is_object($localtoolbox)) {
+            $themesettings = \theme_adaptable\toolbox::get_settings();
+            $retr = $localtoolbox->get_user_settings($themesettings, $this->page, $this).$retr;
         }
-        // Then 'b' must be the enrol entry.
-        return 1;
+        return $retr;
     }
 
     /**
@@ -1632,31 +1615,28 @@ trait core_renderer_toolbox {
         $overridelist = false;
         $overridetype = 'off';
 
-        if (!empty($this->page->theme->settings->navbardisplayicons)) {
-            $navbardisplayicons = true;
-        } else {
-            $navbardisplayicons = false;
+        $themesettings = \theme_adaptable\toolbox::get_settings();
+        $navbardisplayicons = (!empty($themesettings->navbardisplayicons));
+        $navbardisplaytitles = (!empty($themesettings->navbardisplaytitles));
+        if (!$navbardisplaytitles && !$navbardisplayicons) {
+            // Can't have nothing!
+            $navbardisplaytitles = true;
         }
-
-        $mysitesmaxlength = '30';
-        if (!empty($this->page->theme->settings->mysitesmaxlength)) {
-            $mysitesmaxlength = $this->page->theme->settings->mysitesmaxlength;
-        }
-
-        $mysitesmaxlengthhidden = $mysitesmaxlength - 3;
 
         $branchsort = 9998;
 
         if (isloggedin() && !isguestuser()) {
-            if (!empty($this->page->theme->settings->enablehome)) {
+            if (!empty($themesettings->enablehome)) {
                 $branchlabel = '';
                 $branchtitle = get_string('home', 'theme_adaptable');
                 if ($navbardisplayicons) {
                     $branchlabel .= \theme_adaptable\toolbox::getfontawesomemarkup('home', ['fa-lg', 'mr-1']);
                 }
-                $branchlabel .= $branchtitle;
+                if ($navbardisplaytitles) {
+                    $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
+                }
 
-                if (!empty($this->page->theme->settings->enablehomeredirect)) {
+                if (!empty($themesettings->enablehomeredirect)) {
                     $branchurl = new url('/?redirect=0');
                 } else {
                     $branchurl = new url('/');
@@ -1664,99 +1644,70 @@ trait core_renderer_toolbox {
                 $branch = $menu->add($branchlabel, $branchurl, $branchtitle, $branchsort);
             }
 
-            if (!empty($this->page->theme->settings->enablemyhome)) {
+            if (!empty($themesettings->enablemyhome)) {
                 $branchlabel = '';
                 $branchtitle = get_string('myhome');
                 if ($navbardisplayicons) {
                     $branchlabel .= \theme_adaptable\toolbox::getfontawesomemarkup('dashboard', ['fa-lg', 'mr-1']);
                 }
-                $branchlabel .= $branchtitle;
+                if ($navbardisplaytitles) {
+                    $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
+                }
                 $branchurl = new url('/my/index.php');
                 $branchsort++;
                 $branch = $menu->add($branchlabel, $branchurl, $branchtitle, $branchsort);
             }
 
-            if (!empty($this->page->theme->settings->enablemycourses)) {
+            if (!empty($themesettings->enablemycourses)) {
                 $branchlabel = '';
                 $branchtitle = get_string('courses');
                 if ($navbardisplayicons) {
                     $branchlabel .= \theme_adaptable\toolbox::getfontawesomemarkup('th', ['fa-lg', 'mr-1']);
                 }
-                $branchlabel .= $branchtitle;
+                if ($navbardisplaytitles) {
+                    $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
+                }
                 $branchurl = new url('/my/courses.php');
                 $branchsort++;
                 $branch = $menu->add($branchlabel, $branchurl, $branchtitle, $branchsort);
             }
 
-            if (!empty($this->page->theme->settings->enableevents)) {
+            if (!empty($themesettings->enableevents)) {
                 $branchlabel = '';
                 $branchtitle = get_string('events', 'theme_adaptable');
                 if ($navbardisplayicons) {
                     $branchlabel .= \theme_adaptable\toolbox::getfontawesomemarkup('calendar', ['fa-lg', 'mr-1']);
                 }
-                $branchlabel .= $branchtitle;
+                if ($navbardisplaytitles) {
+                    $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
+                }
 
                 $branchurl = new url('/calendar/view.php');
                 $branchsort++;
                 $branch = $menu->add($branchlabel, $branchurl, $branchtitle, $branchsort);
             }
 
-            $overridetype = null;
-            $overridelist = null;
-
-            if (!empty($this->page->theme->settings->mysitessortoverride)) {
-                $overridetype = $this->page->theme->settings->mysitessortoverride;
-            }
-
-            if (!empty($this->page->theme->settings->mysitessortoverridefield)) {
-                $overridelist = $this->page->theme->settings->mysitessortoverridefield;
-            }
-
-            if (($overridetype == 'profilefields' || $overridetype == 'profilefieldscohort') && (isset($overridelist))) {
-                $overridelist = $this->get_profile_field_contents($overridelist);
-
-                if ($overridetype == 'profilefieldscohort') {
-                    $overridelist = array_merge($this->get_cohort_enrollments(), $overridelist);
-                }
-            }
-
-            if ($overridetype == 'strings' && isset($overridelist)) {
-                $overridelist = explode(',', $overridelist);
-            }
-
             $localtoolbox = \theme_adaptable\toolbox::get_local_toolbox();
             if (is_object($localtoolbox)) {
-                $themesettings = \theme_adaptable\toolbox::get_settings();
                 $localtoolbox->get_mycourses(
                     $menu,
                     $branchsort,
-                    $navbardisplayicons,
-                    $overridelist,
-                    $overridetype,
-                    $mysitesmaxlength,
-                    $mysitesmaxlengthhidden,
-                    $this->page->theme->settings,
+                    $themesettings,
                     $this->page,
                     $this
                 );
             }
 
-            if (!empty($this->page->theme->settings->enablethiscourse)) {
+            if (!empty($themesettings->enablethiscourse)) {
                 if (isset($COURSE->id) && $COURSE->id != SITEID) {
-                    $branchlabel = '';
                     $branchtitle = get_string('thiscourse', 'theme_adaptable');
+                    $branchlabel = '';
                     if ($navbardisplayicons) {
                         $branchlabel .=
-                            \theme_adaptable\toolbox::getfontawesomemarkup('sitemap', ['mr-1', 'fa-lg']) . '<span class="menutitle">';
+                            \theme_adaptable\toolbox::getfontawesomemarkup('sitemap', ['mr-1', 'fa-lg']);
                     }
-                    $branchlabel .= $branchtitle;
-                    if ($navbardisplayicons) {
-                        $branchlabel .= '</span>';
-                    }
-
-                    // Check the option of displaying a sub-menu arrow symbol.
-                    if (!empty($this->page->theme->settings->navbardisplaysubmenuarrow)) {
-                        $branchlabel .= \theme_adaptable\toolbox::getfontawesomemarkup('caret-down', ['ml-1']);
+                    if ($navbardisplaytitles) {
+                        $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
                     }
 
                     $branchurl = $this->page->url;
@@ -1764,13 +1715,13 @@ trait core_renderer_toolbox {
                     $branch = $menu->add($branchlabel, $branchurl, $branchtitle, $branchsort);
 
                     // Course sections.
-                    if ($this->page->theme->settings->enablecoursesections) {
+                    if ($themesettings->enablecoursesections) {
                         $this->create_course_sections_menu($branch);
                     }
 
                     // Display Participants.
                     $branchmenusort = 10000;
-                    if ($this->page->theme->settings->displayparticipants) {
+                    if ($themesettings->displayparticipants) {
                         $branchtitle = get_string('people', 'theme_adaptable');
                         $branchlabel = \theme_adaptable\toolbox::getfontawesomemarkup(
                             'users',
@@ -1778,15 +1729,17 @@ trait core_renderer_toolbox {
                             [],
                             '',
                             $branchtitle
-                        ) . $branchtitle;
+                        );
+                        $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
                         $branchurl = new url('/user/index.php', ['id' => $this->page->course->id]);
                         $branch->add($branchlabel, $branchurl, $branchtitle, $branchmenusort);
                     }
 
                     // Display Grades.
-                    if ($this->page->theme->settings->displaygrades) {
+                    if ($themesettings->displaygrades) {
                         $branchtitle = get_string('grades');
-                        $branchlabel = $this->pix_icon('i/grades', $branchtitle, '') . $branchtitle;
+                        $branchlabel = $this->pix_icon('i/grades', $branchtitle, '');
+                        $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
                         $branchurl = new url('/grade/report/index.php', ['id' => $this->page->course->id]);
                         $branchmenusort++;
                         $branch->add($branchlabel, $branchurl, $branchtitle, $branchmenusort);
@@ -1795,7 +1748,8 @@ trait core_renderer_toolbox {
                     // Kaltura video gallery.
                     if (\theme_adaptable\toolbox::kalturaplugininstalled()) {
                         $branchtitle = get_string('nav_mediagallery', 'local_kalturamediagallery');
-                        $branchlabel = $this->pix_icon('media-gallery', $branchtitle, 'local_kalturamediagallery') . $branchtitle;
+                        $branchlabel = $this->pix_icon('media-gallery', $branchtitle, 'local_kalturamediagallery');
+                        $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
                         $branchurl = new url(
                             '/local/kalturamediagallery/index.php',
                             ['courseid' => $this->page->course->id]
@@ -1806,9 +1760,10 @@ trait core_renderer_toolbox {
 
                     // Display Competencies.
                     if (get_config('core_competency', 'enabled')) {
-                        if ($this->page->theme->settings->enablecompetencieslink) {
+                        if ($themesettings->enablecompetencieslink) {
                             $branchtitle = get_string('competencies', 'competency');
-                            $branchlabel = $this->pix_icon('i/competencies', $branchtitle, '') . $branchtitle;
+                            $branchlabel = $this->pix_icon('i/competencies', $branchtitle, '');
+                            $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
                             $branchurl = new url(
                                 '/admin/tool/lp/coursecompetencies.php',
                                 ['courseid' => $this->page->course->id]
@@ -1821,20 +1776,22 @@ trait core_renderer_toolbox {
                     // Display activities.
                     $data = theme_adaptable_get_course_activities();
                     foreach ($data as $modname => $modfullname) {
+                        $branchmenusort++;
+                        $branchlabel = '';
                         if ($modname === 'resources') {
-                            $icon = $this->pix_icon('monologo', get_string('pluginname', 'mod_page'), 'mod_page');
-                            $branchmenusort++;
+                            $branchlabel .= $this->pix_icon('monologo', get_string('pluginname', 'mod_page'), 'mod_page');
+                            $branchlabel .= '<span class="menutitle">' . $modfullname . '</span>';
                             $branch->add(
-                                $icon . $modfullname,
+                                $branchlabel,
                                 new url('/course/resources.php', ['id' => $this->page->course->id]),
                                 $modfullname,
                                 $branchmenusort
                             );
                         } else {
-                            $icon = $this->pix_icon('monologo', get_string('pluginname', 'mod_' . $modname), $modname);
-                            $branchmenusort++;
+                            $branchlabel .= $this->pix_icon('monologo', get_string('pluginname', 'mod_' . $modname), $modname);
+                            $branchlabel .= '<span class="menutitle">' . $modfullname . '</span>';
                             $branch->add(
-                                $icon . $modfullname,
+                                $branchlabel,
                                 new url('/mod/' . $modname . '/index.php', ['id' => $this->page->course->id]),
                                 $modfullname,
                                 $branchmenusort
@@ -1851,14 +1808,14 @@ trait core_renderer_toolbox {
             $helpicon = '';
         }
 
-        if (!empty($this->page->theme->settings->helplinkscount)) {
-            for ($helpcount = 1; $helpcount <= $this->page->theme->settings->helplinkscount; $helpcount++) {
+        if (!empty($themesettings->helplinkscount)) {
+            for ($helpcount = 1; $helpcount <= $themesettings->helplinkscount; $helpcount++) {
                 $enablehelpsetting = 'enablehelp' . $helpcount;
-                if (!empty($this->page->theme->settings->$enablehelpsetting)) {
+                if (!empty($themesettings->$enablehelpsetting)) {
                     $access = true;
                     $helpprofilefieldsetting = 'helpprofilefield' . $helpcount;
-                    if (!empty($this->page->theme->settings->$helpprofilefieldsetting)) {
-                        $fields = explode('=', $this->page->theme->settings->$helpprofilefieldsetting);
+                    if (!empty($themesettings->$helpprofilefieldsetting)) {
+                        $fields = explode('=', $themesettings->$helpprofilefieldsetting);
                         $ftype = $fields[0];
                         $setvalue = $fields[1];
                         if (!$this->check_menu_access($ftype, $setvalue, 'help' . $helpcount)) {
@@ -1868,15 +1825,18 @@ trait core_renderer_toolbox {
 
                     if ($access && !$this->hideinforum()) {
                         $helplinktitlesetting = 'helplinktitle' . $helpcount;
-                        if (empty($this->page->theme->settings->$helplinktitlesetting)) {
+                        if (empty($themesettings->$helplinktitlesetting)) {
                             $branchtitle = get_string('helptitle', 'theme_adaptable', ['number' => $helpcount]);
                         } else {
-                            $branchtitle = $this->page->theme->settings->$helplinktitlesetting;
+                            $branchtitle = $themesettings->$helplinktitlesetting;
                         }
-                        $branchlabel = $helpicon . $branchtitle;
+                        $branchlabel = $helpicon;
+                        if ($navbardisplaytitles) {
+                            $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
+                        }
                         $branchurl = new url(
-                            $this->page->theme->settings->$enablehelpsetting,
-                            ['helptarget' => $this->page->theme->settings->helptarget]
+                            $themesettings->$enablehelpsetting,
+                            ['helptarget' => $themesettings->helptarget]
                         );
 
                         $branchsort++;
@@ -1888,24 +1848,17 @@ trait core_renderer_toolbox {
 
         // Custom menu.
         if ((!empty($CFG->custommenuitems)) &&
-            (empty($this->page->theme->settings->disablecustommenu))) {
+            (empty($themesettings->disablecustommenu))) {
             $custommenutitle = \theme_adaptable\toolbox::get_setting('custommenutitle', 'format_plain');
             $branch = null;
             if (!empty($custommenutitle)) {
                 $branchlabel = '';
                 $branchtitle = $custommenutitle;
                 if ($navbardisplayicons) {
-                    $branchlabel .=
-                        \theme_adaptable\toolbox::getfontawesomemarkup('bars', ['mr-1', 'fa-lg']) . '<span class="menutitle">';
+                    $branchlabel .= \theme_adaptable\toolbox::getfontawesomemarkup('ellipsis', ['mr-1', 'fa-lg']);
                 }
-                $branchlabel .= $branchtitle;
-                if ($navbardisplayicons) {
-                    $branchlabel .= '</span>';
-                }
-
-                // Check the option of displaying a sub-menu arrow symbol.
-                if (!empty($this->page->theme->settings->navbardisplaysubmenuarrow)) {
-                    $branchlabel .= \theme_adaptable\toolbox::getfontawesomemarkup('caret-down', ['ml-1']);
+                if ($navbardisplaytitles) {
+                    $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
                 }
 
                 $branchurl = $this->page->url;
@@ -1951,13 +1904,15 @@ trait core_renderer_toolbox {
 
         if (!empty($sectionsformnenu)) { // Rare but possible!
             $branchtitle = get_string('sections', 'theme_adaptable');
-            $branchlabel = \theme_adaptable\toolbox::getfontawesomemarkup(
+            $branchlabel = '';
+            $branchlabel .= \theme_adaptable\toolbox::getfontawesomemarkup(
                 'list-ol',
                 ['icon', 'fa-lg'],
                 [],
                 '',
                 $branchtitle
-            ) . $branchtitle;
+            );
+            $branchlabel .= '<span class="menutitle">' . $branchtitle . '</span>';
             $branch = $menu->add($branchlabel, null, $branchtitle, 100003);
 
             foreach ($sectionsformnenu as $sectionformenu) {
@@ -1986,7 +1941,27 @@ trait core_renderer_toolbox {
     }
 
     /**
-     * Returns html to render tools menu in main navigation bar
+     * Returns html to render user favourites menu on the main navigation bar.
+     *
+     * @param string $menuid The id to use when creating menu.  Used so this could be called for a nav drawer style display.
+     *
+     *
+     * @return string
+     */
+    public function userfav_menu() {
+        $retval = '';
+        $localtoolbox = \theme_adaptable\toolbox::get_local_toolbox();
+
+        if (is_object($localtoolbox)) {
+            $themesettings = \theme_adaptable\toolbox::get_settings();
+            $retval = $localtoolbox->userfav_menu($themesettings, $this->page, $this);
+        }
+
+        return $retval;
+    }
+
+    /**
+     * Returns html to render tools menu on the main navigation bar.
      *
      * @param string $menuid The id to use when creating menu.  Used so this could be called for a nav drawer style display.
      *
@@ -2209,28 +2184,22 @@ trait core_renderer_toolbox {
         $menus = [];
         $visibility = true;
         $nummenus = 0;
+        $themesettings = \theme_adaptable\toolbox::get_settings();
 
-        if (!empty($this->page->theme->settings->menuuseroverride)) {
+        if (!empty($themesettings->menuuseroverride)) {
             $visibility = $this->check_menu_user_visibility();
         }
 
-        $template->showright = false;
-        if (!empty($this->page->theme->settings->menuslinkright)) {
-            $template->showright = true;
-        }
+        $template->showright = (!empty($themesettings->menuslinkright));
 
-        if (!empty($this->page->theme->settings->menuslinkicon)) {
-            $template->menuslinkicon = $this->page->theme->settings->menuslinkicon;
-        } else {
-            $template->menuslinkicon = 'fa-link';
-        }
+        $template->menuslinkicon = (!empty($themesettings->menuslinkicon)) ? $themesettings->menuslinkicon : 'fa-link';
 
         if ($visibility) {
             if (
-                !empty($this->page->theme->settings->topmenuscount) && !empty($this->page->theme->settings->enablemenus)
-                    && (!$this->page->theme->settings->disablemenuscoursepages || $COURSE->id == 1)
+                !empty($themesettings->topmenuscount) && !empty($themesettings->enablemenus)
+                    && (!$themesettings->disablemenuscoursepages || $COURSE->id == 1)
             ) {
-                $topmenuscount = $this->page->theme->settings->topmenuscount;
+                $topmenuscount = $themesettings->topmenuscount;
 
                 for ($i = 1; $i <= $topmenuscount; $i++) {
                     $menunumber = 'menu' . $i;
@@ -2241,9 +2210,9 @@ trait core_renderer_toolbox {
                     $custommenuitems = '';
                     $access = true;
 
-                    if (empty($this->page->theme->settings->$requirelogin) || isloggedin()) {
-                        if (!empty($this->page->theme->settings->$fieldsetting)) {
-                            $fields = explode('=', $this->page->theme->settings->$fieldsetting);
+                    if (empty($themesettings->$requirelogin) || isloggedin()) {
+                        if (!empty($themesettings->$fieldsetting)) {
+                            $fields = explode('=', $themesettings->$fieldsetting);
                             $ftype = $fields[0];
                             $setvalue = $fields[1];
                             if (!$this->check_menu_access($ftype, $setvalue, $menunumber)) {
@@ -2251,11 +2220,11 @@ trait core_renderer_toolbox {
                             }
                         }
 
-                        if (!empty($this->page->theme->settings->$newmenu) && $access == true) {
+                        if (!empty($themesettings->$newmenu) && $access == true) {
                             $nummenus++;
-                            $menu = ($this->page->theme->settings->$newmenu);
-                            $title = ($this->page->theme->settings->$newmenutitle);
-                            $custommenuitems = $this->parse_custom_menu($menu, format_string($title));
+                            $menu = $themesettings->$newmenu;
+                            $title = format_string($themesettings->$newmenutitle);
+                            $custommenuitems = $this->parse_custom_menu($menu, $title, $title);
                             $custommenu = new custom_menu($custommenuitems, current_language());
                             $menus[] = $this->render_overlay_menu($custommenu);
                         }
@@ -2300,11 +2269,7 @@ trait core_renderer_toolbox {
             }
         }
 
-        if ($showlinktext == false) {
-            $template->showlinktext = false;
-        } else {
-            $template->showlinktext = true;
-        }
+        $template->showlinktext = $showlinktext;
 
         return $this->render_from_template('theme_adaptable/overlaymenu', $template);
     }
@@ -2466,88 +2431,22 @@ trait core_renderer_toolbox {
     }
 
     /**
-     * Returns list of cohort enrollments
-     *
-     * @return array
-     */
-    public function get_cohort_enrollments() {
-        global $DB, $USER;
-        $userscohorts = $DB->get_records('cohort_members', ['userid' => $USER->id]);
-        $courses = [];
-        if ($userscohorts) {
-            $cohortedcourseslist = $DB->get_records_sql('select '
-                    . 'courseid '
-                    . 'from {enrol} '
-                    . 'where enrol = "cohort" '
-                    . 'and customint1 in (?)', array_keys($userscohorts));
-            $cohortedcourses = $DB->get_records_list('course', 'id', array_keys($cohortedcourseslist), null, 'shortname');
-            foreach ($cohortedcourses as $course) {
-                $courses[] = $course->shortname;
-            }
-        }
-        return($courses);
-    }
-
-    /**
-     * Returns contents of multiple comma delimited custom profile fields.
-     *
-     * @param string $profilefields delimited list of fields.
-     * @return array of multiple comma delimited custom profile fields.
-     */
-    public function get_profile_field_contents($profilefields) {
-        global $CFG, $USER;
-        $timestamp = 'currentcoursestime';
-        $list = 'currentcourseslist';
-        $time = time();
-
-        if (isset($USER->theme_adaptable_menus[$timestamp])) {
-            if ($USER->theme_adaptable_menus[$timestamp] >= $time) {
-                if (isset($USER->theme_adaptable_menus[$list])) {
-                    return $USER->theme_adaptable_menus[$list];
-                }
-            }
-        }
-
-        $retval = [];
-
-        require_once($CFG->dirroot . '/user/profile/lib.php');
-        require_once($CFG->dirroot . '/user/lib.php');
-        profile_load_data($USER);
-
-        $fields = explode(',', $profilefields);
-        foreach ($fields as $field) {
-            $field = trim($field);
-            $field = "profile_field_$field";
-            if (isset($USER->$field)) {
-                $vals = explode(',', $USER->$field);
-                foreach ($vals as $value) {
-                    $retval[] = trim($value);
-                }
-            }
-        }
-
-        $USER->theme_adaptable_menus[$list] = $retval;
-        $USER->theme_adaptable_menus[$timestamp] = $time + 1000 * 60 * 3; // Sess TTL.
-
-        return $retval;
-    }
-
-    /**
      * Parses / wraps custom menus in HTML.
      *
      * @param string $menu
      * @param string $label
+     * @param string $title
      * @param string $class
      * @param string $close
      *
      * @return string
      */
-    public function parse_custom_menu($menu, $label, $class = '', $close = '') {
+    public function parse_custom_menu($menu, $label, $title, $class = '', $close = '') {
 
         /* Top level menu option.  No URL added after $close (previously was #).
            Done to fix current jquery / Bootstrap version incompatibility with using #
            in target URLS. Ref: Issue 617 on Adaptable theme issues on Bitbucket. */
-        $custommenuitems = $class . $label . $close . "||" . $label . "\n";
+        $custommenuitems = $class . $label . $close . "||" . $title . "\n";
         $arr = explode("\n", $menu);
 
         // We want to force everything inputted under this menu.
@@ -2560,7 +2459,7 @@ trait core_renderer_toolbox {
     }
 
     /**
-     * Hide tools menu in forum to make room for forum search optoin
+     * Hide help, user favourites and tools menu in a forum to make room for forum search option.
      *
      * @return boolean
      */
@@ -2649,14 +2548,14 @@ trait core_renderer_toolbox {
      *
      * @return string
      */
-    public function render_custom_menu(\custom_menu $menu, $wrappre = '', $wrappost = '', $menuid = '') {
+    public function render_custom_menu(core_custom_menu $menu, $wrappre = '', $wrappost = '', $menuid = '') {
         if (!$menu->has_children()) {
             return '';
         }
 
         $content = '';
         foreach ($menu->get_children() as $item) {
-            if (stristr($menuid, 'drawer')) {
+            if ((!empty($menuid)) && (stristr($menuid, 'drawer'))) {
                 $content .= $this->render_custom_menu_item_drawer($item, 0, $menuid, false);
             } else {
                 $content .= $this->render_custom_menu_item($item, 0, $menuid);