+++ db/upgrade.php
@@ -102,6 +102,17 @@ function xmldb_theme_adaptable_upgrade($oldversion = 0) {
         upgrade_plugin_savepoint(true, 2024032803, 'theme', 'adaptable');
     }
 
+    if ($oldversion < 2024100502) {
+        $value = get_config('theme_adaptable', 'navbardisplaysubmenuarrow');
+        if (!empty($value)) {
+            set_config('navbardisplaymenuarrow', $value, 'theme_adaptable');
+            // Prevent replacement when upgrade has already happened in a version for an older Moodle!
+            unset_config('navbardisplaysubmenuarrow', 'theme_adaptable');
+        }
+
+        upgrade_plugin_savepoint(true, 2024100502, 'theme', 'adaptable');
+    }
+
     // Automatic 'Purge all caches'....
     purge_all_caches();